backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
requests>=2.31
